requirements.txt
sentence_transformers
django==3.2
djangorestframework
pandas
django-cors-headers